Large refactor
Split log into a log and log_abi crate, so that the macro system can be
used in the absence of malloc. Also split out memory into a separate
crate, and mark it as an allocator so rust can make some better choices
about it. Split some CPU utility functions out into constants, maybe not
the best place for it.