WIP do work on fixing allocator
It's broken because I made bad assumptions about total orders

Also, working on removing serde/corepack/the binary format that I had been
working on because I realized that it wasn't that interesting, really, and it
was worse the way that I implemented it. ELF is at least tractable, and it makes
my life easier.